dental_finder/pyproject.toml
[tool.poetry]
name = "dental finder"
version = "0.1.0"
description = ""
authors = ["vfamim <vfamim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.3"
numpy = "^1.26.2"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
